Display '\x2011' (non-breaking hyphen) as a regular hyphen (\x002D)
The non-breaking hyphen was once introduced to be able to make the
distinction between the braille contraction '-' and the braille hyphen
'-'. The distinction was needed because in a post-translation step, '-t'
had to be replaced by '-m', but not if the '-' was representing a
contraction.

Now, because Pipeline2 makes a distinction between SHY (soft hyphen,
which maps to 't') and ZWSP (zero-width space, which maps to 'm'), there
is no need for this post-translation step anymore, and therefore the
'\x2011' is not needed either.