Drop conversion from mod_src to tables
Historically there was a `src` directory where tables were stored and
edited in some pre-cooked form, i.e. the dot patterns were listed as
ascii braille instead of dot patterns. This is now gone with the
introduction of maketable based tables. All that was left really was a
make target that would copy the tables from `src` to `tables`. This is
now gone and with it also the Perl code that did the conversion.

Might be a small step for the sbs braille tables but a big step
towards upstreaming the tables.